+++ CHANGELOG.md
@@ -47,6 +47,9 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - A custom error in the error log when you try adding unknown risk tags to user.
 
 - A custom error in the error log when you try adding a user to a detection list who is already added.
+- Graceful handling of keyboard interrupts (ctrl-c) so stack traces aren't printed to console.
+- Warning message printed when ctrl-c is encountered in the middle of an operation that could cause incorrect checkpoint
+    state, a second ctrl-c is required to quit while that operation is ongoing. 
 
 ### Fixed